Removed unnecessary database code
